requirements.txt
langchain~=0.3.26
langchain_groq~=0.3.2
langchain_openai~=0.3.23
langchain_community~=0.3.24
python-dotenv~=1.1.0
pyyaml~=6.0.2
langgraph~=0.4.8
black~=25.1.0
tqdm~=4.67.1
aiohttp~=3.11.14
bs4>=0.0.2
datasets>=3.6.0
ragas~=0.2.15
rapidfuzz~=3.13.0
scikit-learn>=1.7.0
pandas~=2.3.0
pyjokes~=0.8.3
spacy~=3.8.7
mcp~=1.9.4
langchain_mcp_adapters
pygithub~=2.6.1
langchain_tavily
selenium~=4.34.0
unstructured~=0.18.1
serpapi
pyowm
PyPDF2